refactor: stellar_grid_key was overcrowding 0,0,0
previous algorithm would produce the value 0 for -32. <= coord <= 32.

new implementation doesn't pack the bits quite so tightly, allowing the optimizer to potentially recognize them as uint16s and a uint32. it also produces -1 for -32.<=coord<0., -2 for -64<=coord<-32, etc. ensuring the 'home' sector isn't 3x more densely populated than it needs to be